refactor(theme): move Change Theme into the hamburger menu
The EDIT-layout color-theme control moves out of the crowded EDIT toolbar into a 'Change Theme' item in the app hamburger menu (App section), opening a modal of theme swatches + a custom background image. Removes the toolbar THEME button and EditThemePicker; the theme still applies to the EDIT layout via editThemeStore.
